CNN Model first results
In this commit, I have been able to set up a DGX Account with the IT Office. All development of models will now all be done on the DGX. 
I was able to create a simple CNN architecture which has been detailed in the create_model function. 
Results can be observed in the file