More models ran
In this commit, I have ran more models to see if precision and recall of Normal images will increase. 
In order to do that I used both SGD instead of Adam as it was giving more stable results. The LR was kept the same. 
Furthermore, Bath size was halved to see if more frequent weight updates result in higher recall and precision for Normal images
The next commit will most likely be adding dropout to see if overfitting on Pneuomonia images can be reduced. 

Thanks